+++ monitor/app/staking/metrics.go
@@ -33,25 +33,4 @@ var (
 		Name:      "stake_median_gwei",
 		Help:      "Median stake size in gwei",
 	})
-
-	cChainSupply = promauto.NewGauge(prometheus.GaugeOpts{
-		Namespace: "monitor",
-		Subsystem: "staking",
-		Name:      "supply_cchain_gwei",
-		Help:      "Token supply on the consensus chain in gwei",
-	})
-
-	eChainSupply = promauto.NewGauge(prometheus.GaugeOpts{
-		Namespace: "monitor",
-		Subsystem: "staking",
-		Name:      "total_supply_erc20_gwei",
-		Help:      "ERC20 token supply on Ethereum chain in gwei",
-	})
-
-	bridgeBalance = promauto.NewGauge(prometheus.GaugeOpts{
-		Namespace: "monitor",
-		Subsystem: "staking",
-		Name:      "bridge_balance_gwei",
-		Help:      "ERC20 token balance of the bridge contract on Ethereum in gwei",
-	})
 )
